Add SetFileEndings command with InputHandler
This command adds the `set_line_endings` command with support for the new input
handlers as alternative to the built-in `set_line_ending` which does not yet.

The corresponding entries in the _Command Palette_ are updated.

The entry is now listed as "File: Set Line Endings" right next to
"File: Set Encoding"